requirements.txt
pandas
plotly-express
pyodbc
streamlit